Basic unit test
